requirements.txt
langchain
python-dotenv
langchain_community
langchain-openai
langchain-ollama
langchain-groq
fastapi
uvicorn
langserve
streamlit